Use int instead of short, arithmetic casting is getting verbose

+++ MonoGame WaTor/GameObjects/Fish.cs
@@ -16,7 +16,7 @@ namespace MonoGame_WaTor.GameObjects
         public override Color Color => FishColor;
         public override Texture2D Texture => FishTexture;
 
-        public Fish(EntityGrid world, PriorityGroupedList<Entity> entities, short x, short y, bool addToWorld = true) : base(world, entities, x, y, addToWorld) { }
+        public Fish(EntityGrid world, PriorityGroupedList<Entity> entities, int x, int y, bool addToWorld = true) : base(world, entities, x, y, addToWorld) { }
 
         public static void LoadStaticContent(GraphicsDevice graphics)
         {
@@ -31,6 +31,7 @@ namespace MonoGame_WaTor.GameObjects
 
         public override void Update()
         {
+
         }
     }
 }
